Squash #2137 - Move tests to the correct module test file, rename test file to reflect renamed celery task, clean up imports
